feat: remove mongo backend support for users and threads/comments views This PR removes support for mongo/mysql backend switching logic for users and threads/comments views. Also, update tests to remove mongo support.
