+++ tests/test_views/test_threads.py
@@ -1,13 +1,17 @@
 """Test threads api endpoints."""
 
 import time
+from datetime import datetime
 from typing import Any, Optional
 import pytest
 
-from forum.backends.mongodb.users import Users
+from django.contrib.auth import get_user_model
 from test_utils.client import APIClient
+from forum.backends.mysql.models import ReadState
+from forum.backends.mysql.api import MySQLBackend
 
 pytestmark = pytest.mark.django_db
+User = get_user_model()
 
 
 def setup_models(
@@ -71,9 +75,189 @@ def create_comments_in_a_thread(backend: Any, thread_id: str) -> tuple[str, str]
     return comment_id_1, comment_id_2
 
 
-def test_update_thread(api_client: APIClient, patched_get_backend: Any) -> None:
+def is_thread_id_exists_in_user_read_state(user_id: str, thread_id: str) -> bool:
+    """
+    Return True or False if thread_id exists in read_states of any user using Django ORM.
+    Assumes a User model with a JSONField or related model for read_states.
+    """
+    return ReadState.objects.filter(
+        user_id=user_id, last_read_times__comment_thread=thread_id
+    ).exists()
+
+
+def test_read_states_deletion_of_a_thread_on_thread_deletion(
+    api_client: APIClient,
+    patched_mysql_backend: MySQLBackend,
+) -> None:
+    """Test delete read_states of the thread on deletion of a thread for mongodb."""
+    user_id, thread_id = setup_models(backend=patched_mysql_backend)
+    comment_id_1, comment_id_2 = create_comments_in_a_thread(
+        patched_mysql_backend, thread_id
+    )
+    thread_from_db = patched_mysql_backend.get_thread(thread_id)
+    assert thread_from_db is not None
+    assert thread_from_db["comment_count"] == 2
+    get_thread_response = api_client.get_json(
+        f"/api/v2/threads/{thread_id}",
+        params={
+            "recursive": False,
+            "with_responses": True,
+            "user_id": int(user_id),
+            "mark_as_read": False,
+            "resp_skip": 0,
+            "resp_limit": 10,
+            "reverse_order": "true",
+            "merge_question_type_responses": False,
+        },
+    )  # call get_thread API to save read_states of this thread in user model
+    assert get_thread_response.status_code == 200
+    assert is_thread_id_exists_in_user_read_state(user_id, thread_id) is True
+    response = api_client.delete_json(f"/api/v2/threads/{thread_id}")
+    assert response.status_code == 200
+    assert patched_mysql_backend.get_thread(thread_id) is None
+    assert patched_mysql_backend.get_comment(comment_id_1) is None
+    assert patched_mysql_backend.get_comment(comment_id_2) is None
+    assert (
+        patched_mysql_backend.get_subscription(
+            subscriber_id=user_id, source_id=thread_id
+        )
+        is None
+    )
+    assert is_thread_id_exists_in_user_read_state(user_id, thread_id) is False
+
+
+def test_read_states_deletion_on_thread_deletion_without_read_states(
+    api_client: APIClient,
+    patched_mysql_backend: MySQLBackend,
+) -> None:
+    """Test delete read_states of the thread on deletion of a thread when there are no read states."""
+    user_id, thread_id = setup_models(backend=patched_mysql_backend)
+    comment_id_1, comment_id_2 = create_comments_in_a_thread(
+        patched_mysql_backend, thread_id
+    )
+    thread_from_db = patched_mysql_backend.get_thread(thread_id)
+    assert thread_from_db is not None
+    assert thread_from_db["comment_count"] == 2
+    assert is_thread_id_exists_in_user_read_state(user_id, thread_id) is False
+
+    response = api_client.delete_json(f"/api/v2/threads/{thread_id}")
+    assert response.status_code == 200
+    assert patched_mysql_backend.get_thread(thread_id) is None
+    assert patched_mysql_backend.get_comment(comment_id_1) is None
+    assert patched_mysql_backend.get_comment(comment_id_2) is None
+    assert (
+        patched_mysql_backend.get_subscription(
+            subscriber_id=user_id, source_id=thread_id
+        )
+        is None
+    )
+    assert is_thread_id_exists_in_user_read_state(user_id, thread_id) is False
+
+
+def test_read_states_deletion_on_thread_deletion_with_multiple_read_states(
+    api_client: APIClient,
+    patched_mysql_backend: MySQLBackend,
+) -> None:
+    """Test delete read_states of the thread on deletion of a thread when there are multiple read states."""
+    # Setup first thread and read state
+    user_id_1, thread_id_1 = setup_models(backend=patched_mysql_backend)
+    get_thread_response = api_client.get_json(
+        f"/api/v2/threads/{thread_id_1}",
+        params={
+            "recursive": False,
+            "with_responses": True,
+            "user_id": int(user_id_1),
+            "mark_as_read": True,
+            "resp_skip": 0,
+            "resp_limit": 10,
+            "reverse_order": "true",
+            "merge_question_type_responses": False,
+        },
+    )
+    assert get_thread_response.status_code == 200
+    assert is_thread_id_exists_in_user_read_state(user_id_1, thread_id_1) is True
+
+    # Setup second thread and read state
+    user_id_2, thread_id_2 = setup_models(
+        backend=patched_mysql_backend,
+        user_id="2",
+        username="user2",
+        course_id="course2",
+    )
+    get_thread_response = api_client.get_json(
+        f"/api/v2/threads/{thread_id_2}",
+        params={
+            "recursive": False,
+            "with_responses": True,
+            "user_id": int(user_id_2),
+            "mark_as_read": True,
+            "resp_skip": 0,
+            "resp_limit": 10,
+            "reverse_order": "true",
+            "merge_question_type_responses": False,
+        },
+    )
+    assert get_thread_response.status_code == 200
+    assert is_thread_id_exists_in_user_read_state(user_id_2, thread_id_2) is True
+
+    # Delete first thread and verify its read state is removed while second remains
+    response = api_client.delete_json(f"/api/v2/threads/{thread_id_1}")
+    assert response.status_code == 200
+    assert patched_mysql_backend.get_thread(thread_id_1) is None
+    assert is_thread_id_exists_in_user_read_state(user_id_1, thread_id_1) is False
+    assert is_thread_id_exists_in_user_read_state(user_id_2, thread_id_2) is True
+
+
+def test_read_states_deletion_checks_thread_id_existence(
+    api_client: APIClient,
+    patched_mysql_backend: MySQLBackend,
+) -> None:
+    """Test that read state deletion only occurs when thread_id exists in last_read_times."""
+    user_id, thread_id = setup_models(backend=patched_mysql_backend)
+    _, other_thread_id = setup_models(backend=patched_mysql_backend)
+
+    read_states = [
+        {
+            "course_id": "course1",
+            "last_read_times": {other_thread_id: datetime.now()},
+        }
+    ]
+    patched_mysql_backend.update_user(user_id, {"read_states": read_states})
+
+    assert is_thread_id_exists_in_user_read_state(user_id, other_thread_id) is True
+    assert is_thread_id_exists_in_user_read_state(user_id, thread_id) is False
+
+    response = api_client.delete_json(f"/api/v2/threads/{thread_id}")
+    assert response.status_code == 200
+    assert is_thread_id_exists_in_user_read_state(user_id, other_thread_id) is True
+
+
+def test_filter_flagged_posts(
+    api_client: APIClient, patched_mysql_backend: MySQLBackend
+) -> None:
+    """Test filter flagged posts through get thread API."""
+    backend = patched_mysql_backend
+    user_id, thread_id = setup_models(backend)
+    tests_flags = [(True, "1"), (False, "0")]
+    for flagged, abuse_flaggers in tests_flags:
+        action = "flag" if flagged else "unflag"
+        response = api_client.put_json(
+            path=f"/api/v2/threads/{thread_id}/abuse_{action}",
+            data={"user_id": str(user_id), "count_flagged": True},
+        )
+        params = {"course_id": "course1", "flagged": flagged}
+        response = api_client.get_json("/api/v2/threads", params)
+        assert response.status_code == 200
+        if action == "unflag":
+            assert response.json()["collection"] == []
+        else:
+            result = response.json()["collection"][0]
+            assert result["abuse_flaggers"] == [abuse_flaggers]
+
+
+def test_update_thread(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test updaing a thread."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend=backend)
     response = api_client.put_json(
         f"/api/v2/threads/{thread_id}",
@@ -97,10 +281,10 @@ def test_update_thread(api_client: APIClient, patched_get_backend: Any) -> None:
 
 
 def test_update_thread_without_user_id(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test updaing a thread without user id."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend=backend)
     response = api_client.put_json(
         f"/api/v2/threads/{thread_id}",
@@ -121,9 +305,9 @@ def test_update_thread_without_user_id(
     assert updated_thread_from_db["thread_type"] == "question"
 
 
-def test_update_close_reason(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_update_close_reason(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test close a thread through update thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend=backend)
     response = api_client.put_json(
         f"/api/v2/threads/{thread_id}",
@@ -143,10 +327,10 @@ def test_update_close_reason(api_client: APIClient, patched_get_backend: Any) ->
 
 
 def test_closing_and_reopening_thread_clears_reason_code(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test close a thread and reopen a thread through update thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend=backend)
     response = api_client.put_json(
         f"/api/v2/threads/{thread_id}",
@@ -175,10 +359,10 @@ def test_closing_and_reopening_thread_clears_reason_code(
 
 
 def test_update_thread_not_exist(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test thread does not exists through update thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     wrong_thread_id = backend.generate_id()
     response = api_client.put_json(
         f"/api/v2/threads/{wrong_thread_id}",
@@ -190,9 +374,9 @@ def test_update_thread_not_exist(
     assert response.status_code == 400
 
 
-def test_unicode_data(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_unicode_data(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test data through update thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend=backend)
     texts = ["测试", "テスト", "test"]
     for text in texts:
@@ -212,9 +396,9 @@ def test_unicode_data(api_client: APIClient, patched_get_backend: Any) -> None:
         assert updated_thread_from_db["title"] == text
 
 
-def test_delete_thread(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_delete_thread(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test delete a thread."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend=backend)
     comment_id_1, comment_id_2 = create_comments_in_a_thread(backend, thread_id)
     thread_from_db = backend.get_thread(thread_id)
@@ -229,18 +413,18 @@ def test_delete_thread(api_client: APIClient, patched_get_backend: Any) -> None:
 
 
 def test_delete_thread_not_exist(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test thread does not exists through delete thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     wrong_thread_id = backend.generate_id()
     response = api_client.delete_json(f"/api/v2/threads/{wrong_thread_id}")
     assert response.status_code == 400
 
 
-def test_invalid_data(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_invalid_data(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test invalid data"""
-    setup_models(patched_get_backend)
+    setup_models(patched_mysql_backend)
     response = api_client.get_json("/api/v2/threads", {})
     assert response.status_code == 400
 
@@ -249,9 +433,9 @@ def test_invalid_data(api_client: APIClient, patched_get_backend: Any) -> None:
     assert response.status_code == 400
 
 
-def test_filter_by_course(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_filter_by_course(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test filter threads by course id through get thread API."""
-    setup_models(patched_get_backend)
+    setup_models(patched_mysql_backend)
     params = {"course_id": "course1"}
     response = api_client.get_json("/api/v2/threads", params)
     assert response.status_code == 200
@@ -263,10 +447,10 @@ def test_filter_by_course(api_client: APIClient, patched_get_backend: Any) -> No
 
 
 def test_filter_exclude_standalone(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test filter exclude standalone threads through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     setup_models(backend=backend)
     backend.create_thread(
         {
@@ -293,10 +477,10 @@ def test_filter_exclude_standalone(
 
 
 def test_api_with_count_flagged(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test thread API with count flagged."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     comment_id_1, comment_id_2 = create_comments_in_a_thread(backend, thread_id)
 
@@ -323,9 +507,11 @@ def test_api_with_count_flagged(
     assert results[0]["abuse_flagged_count"] == 2
 
 
-def test_no_matching_course_id(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_no_matching_course_id(
+    api_client: APIClient, patched_mysql_backend: Any
+) -> None:
     """Test no matching course id through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     setup_models(backend)
     wrong_course_id = "abc"
     params = {"course_id": wrong_course_id}
@@ -335,9 +521,9 @@ def test_no_matching_course_id(api_client: APIClient, patched_get_backend: Any) 
     assert len(results) == 0
 
 
-def test_filter_by_author(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_filter_by_author(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test filter threads by author id through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id1, _ = setup_models(backend=backend)
     user_id2, _ = setup_models(backend, "2", "user2", "course2")
 
@@ -361,9 +547,9 @@ def test_filter_by_author(api_client: APIClient, patched_get_backend: Any) -> No
     assert len(result) == 0
 
 
-def test_anonymous_threads(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_anonymous_threads(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test Anonymus threads are only visible to Authors"""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     course_id = "course-1"
     author_id = "1"
     author_username = "author-1"
@@ -408,9 +594,9 @@ def test_anonymous_threads(api_client: APIClient, patched_get_backend: Any) -> N
     assert len(result) == 2
 
 
-def test_unresponded_filter(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_unresponded_filter(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test unresponded filter"""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     create_comments_in_a_thread(backend, thread_id)
     setup_models(backend, "2", "user2")
@@ -424,9 +610,9 @@ def test_unresponded_filter(api_client: APIClient, patched_get_backend: Any) -> 
     assert len(thread) == 1
 
 
-def test_filter_by_post_type(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_filter_by_post_type(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test filter threads by thread_type through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     setup_models(backend=backend)
     setup_models(backend, "2", "user2", "course1")
     username_3 = "user3"
@@ -462,10 +648,10 @@ def test_filter_by_post_type(api_client: APIClient, patched_get_backend: Any) ->
 
 
 def test_filter_unanswered_questions(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test filter unanswered questions through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     course_id = "course1"
     username = "user1"
     user_id_1, thread1 = setup_models(backend, "1", "user1", thread_type="question")
@@ -534,9 +720,9 @@ def test_filter_unanswered_questions(
         assert thread["body"] == "Thread 3"
 
 
-def test_get_thread(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_get_thread(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test get thread by thread_id."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     response = api_client.get_json(
         f"/api/v2/threads/{thread_id}",
@@ -560,10 +746,10 @@ def test_get_thread(api_client: APIClient, patched_get_backend: Any) -> None:
 
 
 def test_computes_endorsed_correctly(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test computes endorsed correctly through get thread API."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     comment_id = backend.create_comment(
         {
@@ -594,10 +780,10 @@ def test_computes_endorsed_correctly(
 
 
 def test_no_children_for_informational_request(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test no children returned from get thread by thread_id API"""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     backend.create_comment(
         {
@@ -626,9 +812,9 @@ def test_no_children_for_informational_request(
     assert "children" not in thread
 
 
-def test_mark_as_read(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_mark_as_read(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test mark as read"""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     response = api_client.get_json(
         f"/api/v2/threads/{thread_id}",
@@ -649,9 +835,11 @@ def test_mark_as_read(api_client: APIClient, patched_get_backend: Any) -> None:
     assert thread["read"] is True
 
 
-def test_thread_with_comments(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_thread_with_comments(
+    api_client: APIClient, patched_mysql_backend: Any
+) -> None:
     """Test children returned from get thread by thread_id API"""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     _, thread_id = setup_models(backend)
     comment_id_1, comment_id_2 = create_comments_in_a_thread(backend, thread_id)
 
@@ -674,13 +862,13 @@ def test_thread_with_comments(api_client: APIClient, patched_get_backend: Any) -
 
 
 def test_endorement_is_none_after_unanswering_a_comment_in_question(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """
     Test endorsement is None by marking it as unanswered.
     when that question was initialy marked as answered.
     """
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend, thread_type="question")
     comment_id = backend.create_comment(
         {
@@ -727,13 +915,13 @@ def test_endorement_is_none_after_unanswering_a_comment_in_question(
 
 
 def test_response_for_thread_type_question(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """
     Test responses for thread_type question.
     It varies according to queryparams.
     """
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend, thread_type="question")
     comment_id1 = backend.create_comment(
         {
@@ -831,23 +1019,13 @@ def test_response_for_thread_type_question(
     assert thread["non_endorsed_resp_total"] == 1
 
 
-def is_thread_id_exists_in_user_read_state(user_id: str, thread_id: str) -> bool:
-    """Return True or False if thread_id exists in read_states of any user."""
-    user = Users().find_one({"_id": user_id})
-    if user:
-        for read_state in user.get("read_states", []):
-            if thread_id in read_state.get("last_read_times", {}):
-                return True
-    return False
-
-
-def test_filter_by_group_id(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_filter_by_group_id(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """
     Filter threads by their group_id. This should return:
     - Threads with the specified group ID.
     - Threads that do not belong to any group (i.e., group_id=None).
     """
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     setup_models(backend=backend)
 
     for i in range(2, 5):
@@ -878,13 +1056,13 @@ def test_filter_by_group_id(api_client: APIClient, patched_get_backend: Any) -> 
     assert results[1]["group_id"] is None
 
 
-def test_filter_by_group_ids(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_filter_by_group_ids(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """
     Filter threads by their group IDs. This should return:
     - Threads with the specified group IDs.
     - Threads that do not belong to any group (i.e., group_id=None).
     """
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     setup_models(backend=backend)
     for i in range(2, 5):
         time.sleep(0.001)
@@ -917,13 +1095,13 @@ def test_filter_by_group_ids(api_client: APIClient, patched_get_backend: Any) ->
 
 
 def test_pagination_in_thread_comments(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """
     Test pagination in the thread comments, including various cases such as
     exact limits, skipping, and when there are fewer comments than the limit.
     """
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id, thread_id = setup_models(backend, thread_type="question")
     comments_count = 15  # Total number of comments
     resp_limit = 10  # Limit per request
