+++ tests/test_views/test_users.py
@@ -36,9 +36,9 @@ def setup_10_threads(author_id: str, author_username: str, backend: Any) -> list
     return ids
 
 
-def test_create_user(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_create_user(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test creating a new user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     response = api_client.post_json(
@@ -51,10 +51,10 @@ def test_create_user(api_client: APIClient, patched_get_backend: Any) -> None:
 
 
 def test_create_user_with_existing_id(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test create user with an existing id."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -68,10 +68,10 @@ def test_create_user_with_existing_id(
 
 
 def test_create_user_with_existing_username(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test create user with an existing username."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -84,9 +84,9 @@ def test_create_user_with_existing_username(
     assert response.status_code == 400
 
 
-def test_update_user(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_update_user(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test updating user information."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     new_username = "new-test-user"
@@ -104,10 +104,10 @@ def test_update_user(api_client: APIClient, patched_get_backend: Any) -> None:
 
 
 def test_update_non_existent_user(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test updating non-existent user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     response = api_client.put_json(
         f"/api/v2/users/{user_id}", data={"username": "new-test-user"}
@@ -116,10 +116,10 @@ def test_update_non_existent_user(
 
 
 def test_update_user_with_conflicting_info(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test updating user with conflicting information."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     conflicting_username = "test-user-2"
@@ -137,9 +137,9 @@ def test_update_user_with_conflicting_info(
     assert response.status_code == 400
 
 
-def test_get_user(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_get_user(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test getting user information."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -153,19 +153,21 @@ def test_get_user(api_client: APIClient, patched_get_backend: Any) -> None:
     assert user["username"] == username
 
 
-def test_get_non_existent_user(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_get_non_existent_user(
+    api_client: APIClient, patched_mysql_backend: Any
+) -> None:
     """Test getting non-existent user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     response = api_client.get(f"/api/v2/users/{user_id}")
     assert response.status_code == 404
 
 
 def test_get_user_with_no_votes(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test getting user with no votes."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -178,9 +180,9 @@ def test_get_user_with_no_votes(
     assert user["upvoted_ids"] == []
 
 
-def test_get_user_with_votes(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_get_user_with_votes(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test getting user with votes."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -215,10 +217,10 @@ def test_get_user_with_votes(api_client: APIClient, patched_get_backend: Any) ->
 
 
 def test_get_active_threads_requires_course_id(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test getting active threads requires course id."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -231,9 +233,9 @@ def test_get_active_threads_requires_course_id(
     assert response.json() == {}
 
 
-def test_get_active_threads(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_get_active_threads(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test getting active threads."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -251,10 +253,10 @@ def test_get_active_threads(api_client: APIClient, patched_get_backend: Any) -> 
 
 
 def test_marks_thread_as_read_for_user(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test marking a thread as read for a user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -296,9 +298,9 @@ def test_marks_thread_as_read_for_user(
     assert read_date[thread_id] >= thread["updated_at"]
 
 
-def test_replaces_username(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_replaces_username(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test replace_username api."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -331,10 +333,10 @@ def test_attempts_to_replace_username_of_non_existent_user(
 
 
 def test_attempts_to_replace_username_and_username_on_content(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test replace_username api with content."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -360,10 +362,10 @@ def test_attempts_to_replace_username_and_username_on_content(
 
 
 def test_attempts_to_replace_username_without_sending_new_username(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test replace_username api without sending new username."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
@@ -378,10 +380,10 @@ def test_attempts_to_replace_username_without_sending_new_username(
 
 
 def test_attempts_to_retire_user_without_sending_retired_username(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test retire user api without sending retired username."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     response = api_client.post_json(
         f"/api/v2/users/{user_id}/retire",
@@ -391,10 +393,10 @@ def test_attempts_to_retire_user_without_sending_retired_username(
 
 
 def test_attempts_to_retire_non_existent_user(
-    api_client: APIClient, patched_get_backend: Any
+    api_client: APIClient, patched_mysql_backend: Any
 ) -> None:
     """Test retire non-existent user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     retired_username = "retired_user_test"
     response = api_client.post_json(
@@ -404,9 +406,9 @@ def test_attempts_to_retire_non_existent_user(
     assert response.status_code == 400
 
 
-def test_retire_user(api_client: APIClient, patched_get_backend: Any) -> None:
+def test_retire_user(api_client: APIClient, patched_mysql_backend: Any) -> None:
     """Test retire user."""
-    backend = patched_get_backend
+    backend = patched_mysql_backend
     user_id = backend.generate_id()
     username = "test-user"
     backend.find_or_create_user(
